[TASK] Move PageSeo persistance to separate namespace
Issue #157
